new exported functions EVALUATE, FIRST-NODE, ALL-NODES
A new function EVALUATE is provided that...
  - automatically parses XPath strings.
  - also accepts sexps in the syntax (xpath (sexp-goes-here)).  (The special
    marker XPATH in the CAR is needed to distinguish literal strings from XPath
    strings.)
  - and accepts pre-compiled closures.  Users who want to use this function
    with a customized environment must pass precompiled closures.

For the first two cases, the environment is defined lexically using the
new macro WITH-NAMESPACES.

The context argument to ALL-NODES can be either...
  - an actual context
  - or any other node understood by the XPath protocol, and will then
    be turned into a trivial context automatically

A compiler macro on EVALUATE allows pre-compilation of closures at fasl
load time.  (Macrolet tricks propagate the information from WITH-NAMESPACES
into the compiler macro.  Thanks to chandler on #lisp for the tip.)

darcs-hash:88c7a7985b99372e14071fb9eb8da0bc97d0b054